Make trigger-expression tagged pointer 64-bit-safe (pointer hygiene, stage 3i)
CExpressionEvaluate reuses the CTrigger* ptrTrigger field as a tagged
pointer: a real trigger, or a small BOOLEAN_* operator sentinel (0..4)
when the EELEMENT_STATE_OPERATOR flag is set. AddExpressionOperator
stores the sentinel as (CTrigger*)u4_operator, and EvaluateExpression
reads it back by casting to uint32 and switching against (uint32)BOOLEAN_*.

On x64 the read casts truncate the pointer (C4311) and the store casts a
32-bit int up to a 64-bit pointer (C4312). It happens to still work
(sentinels are tiny and the switch is guarded by the OPERATOR flag), but
widen both sides to uintptr so the tagged-pointer encoding is correct by
construction and warning-free. BOOLEAN_* are #define integer constants,
so the case labels stay valid constant expressions.

Left line 545 (uint32)*str_exp alone: it narrows a dereferenced char
(the operator character), not a pointer.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings), smoke-launched OK. Ninth slice of the uint32->uintptr
sweep after VirtualMem (ba41823), FastHeap (3825cd3), loaders (43c4b87),
audio (1d560a9), Mesh (0491ee3), MemoryLog (fbf5098), CDArray (16afecc),
CRC/hash (3ba77d6).

+++ jp2_pc/Source/Lib/Trigger/ExpressionEvaluate.cpp
@@ -580,36 +580,36 @@ bool CExpressionEvaluate::EvaluateExpression()
 	{
 		if (aelExpression[u4_i].u4State & EELEMENT_STATE_OPERATOR)
 		{
-			switch ((uint32)aelExpression[u4_i].ptrTrigger)
+			switch ((uintptr)aelExpression[u4_i].ptrTrigger)
 			{
 			// these use bitwise operations because C does not have a logical exclusive or, I
 			// supose it would be identical to a bitwise exclusive or.
-			case (uint32)BOOLEAN_AND:
+			case (uintptr)BOOLEAN_AND:
 			{
 				bool b_temp1 = Pop();
 				bool b_temp2 = Pop();
 				Push(b_temp1 & b_temp2);
 				break;
 			}
-			case (uint32)BOOLEAN_OR:
+			case (uintptr)BOOLEAN_OR:
 			{
 				bool b_temp1 = Pop();
 				bool b_temp2 = Pop();
 				Push(b_temp1 | b_temp2);
 				break;
 			}
-			case (uint32)BOOLEAN_XOR:
+			case (uintptr)BOOLEAN_XOR:
 			{
 				bool b_temp1 = Pop();
 				bool b_temp2 = Pop();
 				Push(b_temp1 ^ b_temp2);
 				break;
 			}
-			case (uint32)BOOLEAN_NOT:
+			case (uintptr)BOOLEAN_NOT:
 				Push(!Pop());
 				break;
 
-			case (uint32)BOOLEAN_EQUAL:
+			case (uintptr)BOOLEAN_EQUAL:
 				Push(Pop()==Pop());
 				break;
